+++ lab_05_03/Point.py
@@ -58,7 +58,7 @@ def draw_line(cnv: tk.Canvas, xs: int, ys: int, edges_arr: List[Tuple[int]], col
 # В ответ на нажатие левой кнопкой мышки отрисовывается точка
 
 
-def touch(x_input: int, y_input: int, cnv: tk.Canvas, tree: ttk.Treeview, ZOOM: int, SIDE_PLACE: int, HEIGHT_PLACE: int,
+def touch(x_input: int, y_input: int, cnv: tk.Canvas, tree: ttk.Treeview, ZOOM: float, SIDE_PLACE: int, HEIGHT_PLACE: int,
           edges_arr: List[Tuple[int]], is_painting: bool, change_coord: bool = True, check_in_table: bool = True) -> None:
     if is_painting:
         mb.showerror('Ошибка!', "Дождитесь конца закраски фигуры!")
