+++ lab_07_01/Point_line.py
@@ -53,7 +53,7 @@ def draw_line(cnv: tk.Canvas, point1: Tuple[int], point2: Tuple[int], tag: str, 
 # надстройка над рисованием отрезка
 
 
-def clever_draw_line(cnv: tk.Canvas, tree: ttk.Treeview, point1: Tuple[int], point2: Tuple[int], ZOOM: int) -> None:
+def clever_draw_line(cnv: tk.Canvas, tree: ttk.Treeview, point1: Tuple[int], point2: Tuple[int], ZOOM: float) -> None:
     x1_table, y1_table = point1
     x2_table, y2_table = point2
     x1_input, y1_input, x2_input, y2_input = x1_table * \
@@ -95,7 +95,7 @@ def dist(point1: Tuple[int], point2: Tuple[int]) -> float:
 # рисует отрезок (видимую и невидимую его часть)
 
 
-def Draw_visibl_line(cnv: tk.Canvas, res_points: Tuple[Tuple[int]], points: Tuple[Tuple[int]], is_sth_visibl: bool, color_vis: str, color_unvis: str, ZOOM: int) -> None:
+def Draw_visibl_line(cnv: tk.Canvas, res_points: Tuple[Tuple[int]], points: Tuple[Tuple[int]], is_sth_visibl: bool, color_vis: str, color_unvis: str, ZOOM: float) -> None:
     P1, P2 = points
     P1_z, P2_z = (P1[c.X_PART] * ZOOM, P1[c.Y_PART] *
                   ZOOM), (P2[c.X_PART] * ZOOM, P2[c.Y_PART] * ZOOM)
@@ -114,7 +114,7 @@ def Draw_visibl_line(cnv: tk.Canvas, res_points: Tuple[Tuple[int]], points: Tupl
 # рисование отсекателя
 
 
-def draw_clipper(cnv: tk.Canvas, clipper: List[int], color_clipper: str, ZOOM: int) -> None:
+def draw_clipper(cnv: tk.Canvas, clipper: List[int], color_clipper: str, ZOOM: float) -> None:
     xl, xr = clipper[c.X_LEFT] * ZOOM, clipper[c.X_RIGHT] * ZOOM
     yu, yd = clipper[c.Y_UP] * ZOOM, clipper[c.Y_DOWN] * ZOOM
     points = [(xl, yu), (xl, yd), (xr, yd), (xr, yu)]
